PoC: Do not use shell when running a single test
Run test command directly (without shell). This way it's not necessary to escapy any of the arguments,  even if they contain spaces or other special characters.

However, this breaks backwards compatibility for some setups: If `rspecCommand` is set to contain environment variables or shell builtins it will no longer work.

This is a possible fix for https://github.com/connorshea/vscode-ruby-test-adapter/issues/41

It’s meant as a proof of concept. Therefore the code is deliberately incomplete (it covers `runSingleTest`  only, but `runTestFile` etc. would need to be changed as well).